[services][mcp] encapsulate core of agent as mcp
Prior to these changes, we had a really nice agent in the newsletter
service. It has web search, vector search, and it is a nice little agent
that can be used to find recipes.

This commit encapsulates the core of the agent as an MCP server.

Now we can build out a multi-agent workflow which utilizes the MCP
server at its core. So, not just a newsletter agent, but a meal planning
agent, recipe-generation agent, etc...